+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Change Log
 
+## 1.12.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @digdir/designsystemet-types@1.12.1
+
+## 1.12.0
+
+### Patch Changes
+
+- update npm non-major dependencies ([#4517](https://github.com/digdir/designsystemet/pull/4517))
+
+- update npm non-major dependencies ([#4497](https://github.com/digdir/designsystemet/pull/4497))
+
+- Updated dependencies []:
+  - @digdir/designsystemet-types@1.12.0
+
 ## 1.11.1
 
 ### Patch Changes