+++ packages/types/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @digdir/designsystemet-types
 
+## 1.12.1
+
+## 1.12.0
+
 ## 1.11.1
 
 ## 1.11.0